% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{conv_thresh}
\alias{conv_thresh}
\title{Example 95 Percent Convergence Threshold}
\format{
## `conv_thresh`
A data frame with 1 rows and 5 columns:
\describe{
 \item{fish_type}{Generally a strain, run, and species of fish (e.g.
  Nimbus Fall Chinook = Fall-run Chinook Salmon from Nimbus Hatchery)}
 \item{min_val}{The minimum rSSR value}
 \item{max_val}{The maximum rSSR value}
 \item{threshold}{the rSSR value which represents the `thresh_level` cutoff
  for estimating convergence}
 \item{thresh_level}{The desired convergence level (100-x)}
}
}
\usage{
conv_thresh
}
\description{
Example output from the `conv_thresholds()` function, calculating the 95%
convergence thresholds for the rSSR data found in `ex_rSSR`.
}
\keyword{datasets}
