% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{dat_filt1}
\alias{dat_filt1}
\title{Prefiltered detection data}
\format{
## `dat_filt1`
A data frame with 47,931 rows and 4 columns:
\describe{
  \item{ReceiverSN}{The serial number of the detecting receiver}
  \item{DateTime_Local}{the local time of the detection (tz = America/Los_Angeles)}
  \item{Tag_Code}{The hexadecimal acoustic tag ID code}
  \item{CheckMBP}{A calculated field from the prefilter checking the
   time between acoustic transmissions from the same tag was >0.3secs}
}
}
\source{
Data collected by the California Department of Water Resources 2021
}
\usage{
dat_filt1
}
\description{
An example dataset of real acoustic telemetry detections of fish at several
receivers within the California Central Valley from 2021. These detections
have already been been processed using `prefilter()` from this package or
companion package `filteRjsats`.
}
\keyword{datasets}
