\name{drawVectors}
\alias{drawVectors}
\title{plot the bearings for a set of data}
\usage{
  drawVectors(f, data, subset = TRUE, length, ...)
}
\arguments{
  \item{f}{a formula (with no LHS) specifying the bearing
  and groupings}

  \item{data}{a data frame or spatial data frame with the
  coordinates, bearings, and groupings}

  \item{subset}{logical vector to subset the data}

  \item{length}{length of line segments to draw}

  \item{...}{extra parameters passed to the plot method}
}
\description{
  plot the bearings for a set of data
}

