% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{ex_rSSR}
\alias{ex_rSSR}
\title{Example Renormalized Sum of Squares}
\format{
## `ex_rSSR`
A data frame with 100 rows and 5 columns:
\describe{
 \item{fish_type}{Generally a strain, run, and species of fish (e.g.
  Nimbus Fall Chinook = Fall-run Chinook Salmon from Nimbus Hatchery)}
 \item{mbp_n}{The maximum blanking period (in seconds) used to create a
  set of events}
 \item{SSR}{The sum of squared residuals between this `mbp_n` and the next}
 \item{n}{the total number of events created with this `mbp_n`}
 \item{rSSR}{the renormalized sum of squared residuals between this `mbp_n`
  and the next}
}
}
\usage{
ex_rSSR
}
\description{
Example output from the `renorm_SSR()` function, calculating the
renormalized sum of squares for the "survival" data found in `time_test`.
}
\keyword{datasets}
