% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{filtered_detections}
\alias{filtered_detections}
\title{Previously filtered detection data}
\format{
## `filtered_detections`
A data frame with 7,240 rows and 60 columns:
\describe{
  \item{serial}{The serial number of the detecting receiver}
  \item{local_time}{the local time of the detection (tz = America/Los_Angeles)}
  \item{tag_id}{The hexadecimal acoustic tag ID code}
}
}
\source{
Data collected by the California Department of Water Resources 2021
}
\usage{
filtered_detections
}
\description{
An example dataset of real acoustic telemetry detections of fish at several
receivers within the California Central Valley from 2021. These detections
have already been been processed to remove false positives using the various
filtering functions in this package and in companion package `filteRjsats`.
}
\keyword{datasets}
