% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{fish}
\alias{fish}
\title{Fish Data}
\format{
## `fish`
A data frame with 7,240 rows and 60 columns:
\describe{
  \item{fish_type}{Generally a strain, run, and species of fish (e.g.
  Nimbus Fall Chinook = Fall-run Chinook Salmon from Nimbus Hatchery)}
  \item{TagCode}{The hexadecimal code of the implanted acoustic tag}
  \item{Release_Date}{The release date and time of the fish}
  \item{release_location}{The coded name of the release site}
  \item{length}{The length of the fish in millimeters}
  \item{weight}{The weight of the fish in grams}
  \item{tag_weight}{The weight of the implanted acoustic tag}
  \item{tag_model}{The model number of the implanted acoustic tag}
  \item{PRI}{The pulse rate interval (time between transmissions) of the
  implanted tag, as reported by the manufacturer}
  \item{TagLife}{The expected number of days the tag should continue to
  transmit, as reported by the manufacturer}
}
}
\source{
<https://oceanview.pfeg.noaa.gov/CalFishTrack/pageRealtime_download.html>
}
\usage{
fish
}
\description{
An example dataset of real fish tagged with acoustic telemetry tags and
released within the California Central Valley in 2021 and 2022.
}
\keyword{datasets}
