% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimal_mbp.R
\name{opt_mbp}
\alias{opt_mbp}
\title{Idenitfy the Optimum MBP based on Convergence Threshold}
\usage{
opt_mbp(rSSR_df, thresh_values)
}
\arguments{
\item{rSSR_df}{a dataframe created by created by renorm_SSRduration compare
showing the renormalized sum of squares of the residuals between one
potential blanking period and the next.}

\item{thresh_values}{a dataframe created by conv_thresholds corresponding to
the chosen convergence thresholds.}
}
\value{
A dataframe showing the convergence value and corresponding optimal
maximum blanking period for each grouping.
}
\description{
Takes dataframes created by `renorm_SSR()` and `conv_thresholds()` and
determines the corresponding "optimum" maximum blanking period (MBP) for each
convergence threshold.
}
\examples{
# Determine the optimum mbp
opt_mbp(rSSR_df = ex_rSSR,
        thresh_values = conv_thresh)
}
