% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{raw_detections}
\alias{raw_detections}
\title{Unfiltered detection data}
\format{
## `raw_detections`
A data frame with 55,736 rows and 3 columns:
\describe{
  \item{serial}{The serial number of the detecting receiver}
  \item{local_time}{the local time of the detection (tz = America/Los_Angeles)}
  \item{tag_id}{The hexadecimal acoustic tag ID code}
}
}
\source{
Data collected by the California Department of Water Resources 2021
}
\usage{
raw_detections
}
\description{
An example dataset of real acoustic telemetry detections of fish at several
receivers within the California Central Valley from 2021. These detections
have not been processed to remove false positives.
}
\keyword{datasets}
