% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{reftags}
\alias{reftags}
\title{Example reference tags}
\format{
A vector of example reference tag codes
}
\usage{
reftags
}
\description{
A vector of example reference tag codes
}
\keyword{datasets}
