\docType{package}
\name{telemetr-package}
\alias{telemetr}
\alias{telemetr-package}
\title{Location finding by directional telemetry}
\description{
  Location finding by direction telemetry
}
\details{
  Using bearings and locations to find things.
}
\author{
  \email{B.Rowlingson@gmail.com}
}
\keyword{package}

