% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{time_test}
\alias{time_test}
\title{Example Time Tests for Multi-blanked Detection Data}
\format{
## `time_test`
A data frame with 333,400 rows and 4 columns:
\describe{
 \item{t}{The time (in seconds) against which the duration was compared}
 \item{fish_type}{Generally a strain, run, and species of fish (e.g.
  Nimbus Fall Chinook = Fall-run Chinook Salmon from Nimbus Hatchery)}
 \item{mbp_n}{The maximum blanking period (in seconds) used to create a
  set of events}
 \item{prop_res}{The proportion of all events created with `mbp_n` which have
  a duration longer than time `t`.}
}
}
\usage{
time_test
}
\description{
Example output from the `duration_compare()` function, testing the duration
of detection events found in `blanked_detects`.
}
\keyword{datasets}
