\name{trihub}
\alias{trihub}
\title{triangulate by hub method}
\usage{
  trihub(xytower, bearing, ijob = 1, kappa = 1, sd = 2.5)
}
\arguments{
  \item{xytower}{2 column matrix of xy coords}

  \item{az}{azimuth angles in degrees}

  \item{ijob}{whether to fit kappa}

  \item{kappa}{value of kappa}

  \item{sd}{value of sd}
}
\description{
  triangulate by hub method
}

